Safari-609.1.11.1/LayoutTests/editing/selection/ios/avoid-showing-callout-menu-over-controls-expected.txt

              

Verifies that we don't show the system callout menu over controls. To test manually: focus the editable document in the iframe, select some text, and confirm that the callout bar is shown beneath the selection instead of above the selection.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS centerY(grabberRect) < centerY(menuRect) is true
PASS successfullyParsed is true

TEST COMPLETE
